refactor(web): consolidate to single open-web tool, remove #650 dup, add agent guidance
After both #682 (engine) and #650 (old httpx/DDG web_search) merged to main,
make the insane-search-derived engine the SOLE web-search implementation:

- Remove #650's orphan tool (tools/web/search.py — defined web_search but no
  agent imported it), its tests (test_web_search.py), its docs
  (docs/tools/web-search.md), and its misplaced CHANGELOG bullet.
- Rename tools/web/open_web.py → tools/web/search.py (standard module name,
  freed by the removal); update all importers (tools.py, recon, osint_operator,
  phisher, analyst, tests) and re-export web_fetch from the package __init__.
- Wire web_search/web_fetch into phisher + analyst (per the agent-selection
  review; recon/osint already had them).
- Add 'when to use the web tools' guidance to the recon / osint_operator /
  phisher / analyst system prompts (search→fetch flow, OSINT vs target-gated,
  do NOT point web_fetch at the target's own infra).

Verified: 217 unit tests pass (web tools + roe + sandbox_web engine), ruff
clean, basedpyright 0 errors, bias-check clean. Engine (sandbox_web) unchanged.
